# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Training Analytics API
  version: 1.0.0
paths:
...
  /strava/callback/{username}:
    get:
      tags:
        - Strava
      operationId: stravaCallback
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Authentication successful
          content:
            text/plain:
              schema:
                type: string
...